pyproject.toml
[project]
name = "hasaki-crawl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "beautifulsoup4==4.12.3",
    "datetime>=6.0",
    "logger>=1.4",
    "loguru==0.7.2",
    "lxml==5.1.0",
    "pandas==2.1.4",
    "python-dotenv==1.0.0",
    "ratelimit==2.2.1",
    "requests==2.31.0",
    "robotexclusionrulesparser==1.7.1",
    "selenium==4.16.0",
    "supabase==2.23.0",
    "tenacity==8.2.3",
    "ujson==5.9.0",
    "urllib3==2.1.0",
    "webdriver-manager==4.0.1",
    "websockets==15.0.1",
]
